+++ SSSP/api/schemas/schema_score.py
@@ -2,7 +2,7 @@ from pydantic import BaseModel
 
 class ScoreResponse(BaseModel):
     username: str
-    total_score: float
+    total_score: int
 
     class Config:
         from_attributes=True
